nix: split up derivations into separate .nix files
Splitting up the derivation is a bit more standardized with what people
expect when using Nix, and allows greater future customization, e.g.
specific tools in `shell.nix` that are not build-critical, and more
important jobs in `release.nix`. This also works a bit nicer after some
trials with a CI system and incorporating the default.nix file from a
remote build.

With this change, the provided `default.nix` otherwise looks exactly as
it would upstream: namely, it expects its dependencies as input
parmeters to the derivation, and needs them to be provided.

There is no user-facing change for running `nix-shell` and doing
incremental development, but for `nix-build` and `nix-env`, users must
specify the release.nix file instead now:

    [a@ganon]$ nix-build release.nix -A bcachefs-tools

    [a@ganon]$ nix-env -f release.nix -iA bcachefs-tools

This release.nix file specifics a locked version of nixpkgs, just as it
did before.

The upshot of this is that attribute names are a bit more standardized
in release.nix, and default.nix is completely standardized. A result of
this is that `callPackage` should work fine with the provided
default.nix file, allowing you to easily use fetch targets like fetchgit
to incorporate bcachefs-tools into a build using your own nixpkgs set in
a convenient way:

    let
      my-bcachefs-tools = pkgs.callPackage "${pkgs.fetchFromGitHub {
        owner  = "thoughtpolice";
        repo   = "bcachefs-tools";
        rev    = "...";
        sha256 = "...";
      }/default.nix" { });
    in
    ...
      environment.systemPackages = [ my-bcachefs-tools ];

Signed-off-by: Austin Seipp <aseipp@pobox.com>